Player is a model, models now tracked in a list
The player cube is no longer a basic cube but a model with a mesh, just
like the plane. As such, it responds to the lighting from the shader.
Both models are stored in an array, which can be used to call upon all
models iteratively to do things like unload them or assign the shader.